requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
actuator-msgs==0.0.1
addict==2.4.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
builtin-interfaces==1.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
composition-interfaces==1.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
control-msgs==4.8.0
controller-manager==2.51.0
controller-manager-msgs==2.51.0
cv-bridge==3.2.1
cycler==0.12.1
dash==3.2.0
decorator==5.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
fonttools==4.60.1
gazebo-msgs==3.9.0
generate-parameter-library-py==0.5.0
geometry-msgs==4.9.0
geopandas==1.1.1
gps-msgs==2.1.1
idna==3.11
image-geometry==3.2.1
importlib_metadata==8.7.0
interactive-markers==2.3.2
ipython==8.37.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
laser-geometry==2.4.0
laspy==2.6.1
launch==1.0.10
launch-ros==0.19.10
launch-testing==1.0.10
launch-testing-ros==0.19.10
launch-xml==1.0.10
launch-yaml==1.0.10
lazrs==0.7.0
lifecycle-msgs==1.2.2
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
message-filters==4.3.8
narwhals==2.11.0
nav-msgs==4.9.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.6
open3d==0.19.0
opencv-python==4.12.0.88
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.3.3
parso==0.8.5
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.4.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
pytz==2025.2
PyYAML==6.0.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.6
referencing==0.37.0
requests==2.32.5
resource-retriever==3.1.3
retrying==1.4.2
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.20
ros2action==0.18.13
ros2bag==0.15.15
ros2cli==0.18.13
ros2component==0.18.13
ros2doctor==0.18.13
ros2interface==0.18.13
ros2launch==0.19.10
ros2lifecycle==0.18.13
ros2multicast==0.18.13
ros2node==0.18.13
ros2param==0.18.13
ros2pkg==0.18.13
ros2run==0.18.13
ros2service==0.18.13
ros2topic==0.18.13
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpds-py==0.28.0
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scikit-learn==1.7.2
scipy==1.15.3
scripts==3.9.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
shapely==2.1.2
six==1.17.0
sros2==0.10.7
stack-data==0.6.3
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.15
tf2-kdl==0.25.15
tf2-msgs==0.25.15
tf2-py==0.25.15
tf2-ros-py==0.25.15
tf2-sensor-msgs==0.25.15
tf2-tools==0.25.15
theora-image-transport==2.5.3
threadpoolctl==3.6.0
topic-monitor==0.20.5
tqdm==4.67.1
traitlets==5.14.3
trajectory-msgs==4.9.0
turtlesim==1.4.2
typing_extensions==4.15.0
tzdata==2025.2
unique-identifier-msgs==2.2.1
urllib3==2.5.0
vision-msgs==4.1.1
visualization-msgs==4.9.0
wcwidth==0.2.14
Werkzeug==3.1.3
widgetsnbextension==4.0.15
xacro==2.0.13
zipp==3.23.0
